+++ src/main/java/com/group3979/badmintonbookingbe/model/request/RushHourRequest.java
@@ -8,5 +8,5 @@ import lombok.Setter;
 public class RushHourRequest {
     private int startTime;
     private int endTime;
-    private float rushPrice;
+    private double rushPrice;
 }
